Gates to controll the assignment of the tags

+++ app/Http/Controllers/hobbyTagController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Tag;
 use App\Hobby;
+use Illuminate\Support\Facades\Gate;
 
 class hobbyTagController extends Controller
 {
@@ -25,6 +26,13 @@ class hobbyTagController extends Controller
 
     public function attachTag($hobby_id, $tag_id){
         $hobby = Hobby::find($hobby_id);
+
+        // Bổ sung thêm để thực thi hàm Gate::define trong file Providers/AuthServiceProvider.php
+        if (Gate::denies('connect_hobbyTag', $hobby)) {
+            abort(403, "No no no, this hobby is not yours!");
+        }
+        //End
+
         $tag = Tag::find($tag_id);
         $hobby->tags()->attach($tag_id);
         return back()->with([
@@ -34,6 +42,13 @@ class hobbyTagController extends Controller
 
     public function detachTag($hobby_id, $tag_id){
         $hobby = Hobby::find($hobby_id);
+
+        // Bổ sung thêm để thực thi hàm Gate::define trong file Providers/AuthServiceProvider.php
+        if (Gate::denies('connect_hobbyTag', $hobby)) {
+            abort(403, "No no no, this hobby is not yours!");
+        }
+        //End
+
         $tag = Tag::find($tag_id);
         $hobby->tags()->detach($tag_id);
         return back()->with([
